Eliminated bugs related to renaming.
