+++ apps/pm-user/src/app/app.js
@@ -7,7 +7,9 @@ export function App() {
   return (
     <>
       <LoginForm />
-      <Skeleton />
+      <Skeleton variant="text" />
+      <Skeleton variant="circular" width={40} height={40} />
+      <Skeleton variant="rectangular" width={210} height={118} />
       <span>{output}</span>
     </>
   );
