+++ apps/pm-user/src/app/app.spec.js
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import App from './app';
+import App from './App';
 describe('App', () => {
   it('should render successfully', () => {
     const { baseElement } = render(
